+++ Package.swift
@@ -23,7 +23,7 @@ let package = Package(
     ],
     dependencies: [
         // SpecificationCore: Platform-independent core functionality
-        .package(url: "https://github.com/SoundBlaster/SpecificationCore", from: "1.0.0"),
+        .package(url: "https://github.com/SoundBlaster/SpecificationCore", from: "1.0.1"),
         // Depend on the latest Swift Syntax package for macro support.
         .package(url: "https://github.com/swiftlang/swift-syntax", from: "510.0.0"),
         // Add swift-macro-testing for a simplified macro testing experience.